fix: compiler errors
