+++ Cargo.toml
@@ -23,7 +23,6 @@ serde_json = { version = "1", features = ["alloc"], default_features = false }
 radix_trie = "0.2"
 rustc-hash = "1.1.0"
 miette = { version = "5.7.0", features = ["fancy"] }
-thiserror = "1.0.38"
 dirs = "5.0.1"
 
 # kanata-keyberon = "0.17.0"
@@ -31,6 +30,8 @@ dirs = "5.0.1"
 # Otherwise any changes to the local files will not reflect in the compiled
 # binary.
 kanata-keyberon = { path = "keyberon" }
+
+kanata-parser = { path = "parser" }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 evdev = "=0.12.0"
