+++ parser/Cargo.toml
@@ -11,16 +11,22 @@ license = "LGPL-3.0"
 edition = "2021"
 
 [dependencies]
-clap = { version = "4.1.6", features = [ "std", "derive", "help", "suggestions" ], default_features = false }
 log = { version = "0.4.8", default_features = false }
-simplelog = "0.12.0"
 anyhow = "1"
 parking_lot = "0.12"
 once_cell = "1"
-serde = { version = "1", features = ["alloc", "derive"], default_features = false }
-serde_json = { version = "1", features = ["alloc"], default_features = false }
 radix_trie = "0.2"
 rustc-hash = "1.1.0"
 miette = { version = "5.7.0", features = ["fancy"] }
 thiserror = "1.0.38"
-kanata-keyberon = { path = "keyberon" }
+
+kanata-keyberon = { path = "../keyberon" }
+
+[features]
+cmd = []
+interception_driver = []
+
+[target.'cfg(target_os = "linux")'.dependencies]
+evdev = "=0.12.0"
+
+[target.'cfg(target_os = "windows")'.dependencies]
