+++ parser/src/lib.rs
@@ -1,3 +1,8 @@
 //! A parser for configuration language of [kanata](https://github.com/jtroo/kanata), a keyboard remapper.
 
-// pub mod alloc;
+pub mod cfg;
+pub mod custom_action;
+pub mod keys;
+pub mod layers;
+pub mod sequences;
+pub mod trie;
